Split classes response_options and delivery_response_options
Put Delivery Response Options class in its own file
