+++ src/response_options.cpp
@@ -20,21 +20,5 @@ response_options::response_options(const token_ptr& tok)
 }
 
 /////////////////////////////////////////////////////////////////////////////
-
-delivery_response_options::delivery_response_options()
-		: opts_(MQTTAsync_responseOptions_initializer)
-{
-	opts_.onSuccess = &delivery_token::on_success;
-	opts_.onFailure = &delivery_token::on_failure;
-}
-
-delivery_response_options::delivery_response_options(const delivery_token_ptr& tok)
-		: delivery_response_options()
-{
-	set_token(tok);
-}
-
-/////////////////////////////////////////////////////////////////////////////
 // end namespace 'mqtt'
 }
-
